requirements.txt
azure-ai-projects>=1.0.0
azure-ai-agents>=1.0.0
azure-identity>=1.15.0
openai>=1.0.0
python-dotenv>=1.0.0
